requirements.txt
flask==3.0.2
python-dotenv==1.0.1
Flask-SQLAlchemy==3.1.1
Pymysql==1.1.0
WTForms-SQLAlchemy==0.4.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-mail==0.9.1
numpy==1.26.4